chore: refactoring configs
