+++ src/cores/config/config-service.spec.ts
@@ -8,10 +8,10 @@
  *
  * @copyright (C) lemoncloud.io 2019 - All Rights Reserved.
  */
-import { $engine, _log, _inf, _err, $U } from '../engine/';
-import { credentials, loadJsonSync } from '../tools/';
-import { AWSKMSService } from './aws/aws-kms-service';
-import { expect2, _it } from '../common/test-helper';
+import { $engine, _log, _inf, _err, $U } from '../../engine/';
+import { credentials, loadJsonSync } from '../../tools/';
+import { AWSKMSService } from './../aws/aws-kms-service';
+import { expect2, _it } from '../../common/test-helper';
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////
 //! main test body.
