+++ src/cores/lambda/lambda-web-handler.ts
@@ -18,11 +18,11 @@ import { _log, _inf, _err, $U, $_ } from '../../engine/';
 const NS = $U.NS('HWEB', 'yellow'); // NAMESPACE TO BE PRINTED.
 import { doReportError } from '../../engine/';
 
-import { NextDecoder, NextHandler, NextContext, NextMode, NextIdentityCognito, ProtocolParam } from './../core-types';
+import { NextDecoder, NextHandler, NextContext, NextMode, NextIdentityCognito, ProtocolParam } from './../';
 import { APIGatewayProxyResult, APIGatewayEventRequestContext, APIGatewayProxyEvent } from 'aws-lambda';
 import $lambda, { LambdaHandler, WEBHandler, LambdaHandlerService, Context } from './lambda-handler';
 import $protocol from './../protocol-service';
-import { MyConfigService, ConfigService } from './../config-service';
+import { ConfigService, MyConfigService } from './../config/';
 
 /** ********************************************************************************************************************
  *  COMMON Functions.
